bug fixing , code cleanup

+++ src/main/java/cn/nukkit/block/RedstoneOre.java
@@ -49,7 +49,7 @@ public class RedstoneOre extends Solid {
     @Override
     public int onUpdate(int type) {
         if (type == Level.BLOCK_UPDATE_NORMAL || type == Level.BLOCK_UPDATE_TOUCH) {
-            this.getLevel().setBlock(this, Block.get(Item.GLOWING_REDSTONE_ORE, (int) this.meta), false, true);
+            this.getLevel().setBlock(this, Block.get(Item.GLOWING_REDSTONE_ORE, this.meta), false, true);
 
             return Level.BLOCK_UPDATE_WEAK;
         }
