+++ src/main/java/cn/nukkit/block/Slab.java
@@ -83,11 +83,11 @@ public class Slab extends Transparent {
         this.meta &= 0x07;
         if (face == 0) {
             if (target.getId() == SLAB && (target.getDamage() & 0x08) == 0x08 && (target.getDamage() & 0x07) == (this.meta & 0x07)) {
-                this.getLevel().setBlock(target, Block.get(Item.DOUBLE_SLAB, (int) this.meta), true);
+                this.getLevel().setBlock(target, Block.get(Item.DOUBLE_SLAB, this.meta), true);
 
                 return true;
             } else if (block.getId() == SLAB && (block.getDamage() & 0x07) == (this.meta & 0x07)) {
-                this.getLevel().setBlock(block, Block.get(Item.DOUBLE_SLAB, (int) this.meta), true);
+                this.getLevel().setBlock(block, Block.get(Item.DOUBLE_SLAB, this.meta), true);
 
                 return true;
             } else {
@@ -95,11 +95,11 @@ public class Slab extends Transparent {
             }
         } else if (face == 1) {
             if (target.getId() == SLAB && (target.getDamage() & 0x08) == 0 && (target.getDamage() & 0x07) == (this.meta & 0x07)) {
-                this.getLevel().setBlock(target, Block.get(Item.DOUBLE_SLAB, (int) this.meta), true);
+                this.getLevel().setBlock(target, Block.get(Item.DOUBLE_SLAB, this.meta), true);
 
                 return true;
             } else if (block.getId() == SLAB && (block.getDamage() & 0x07) == (this.meta & 0x07)) {
-                this.getLevel().setBlock(block, Block.get(Item.DOUBLE_SLAB, (int) this.meta), true);
+                this.getLevel().setBlock(block, Block.get(Item.DOUBLE_SLAB, this.meta), true);
 
                 return true;
             }
@@ -107,7 +107,7 @@ public class Slab extends Transparent {
         } else {
             if (block.getId() == SLAB) {
                 if ((block.getDamage() & 0x07) == (this.meta & 0x07)) {
-                    this.getLevel().setBlock(block, Block.get(Item.DOUBLE_SLAB, (int) this.meta), true);
+                    this.getLevel().setBlock(block, Block.get(Item.DOUBLE_SLAB, this.meta), true);
 
                     return true;
                 }
