+++ src/main/java/cn/nukkit/nbt/NBTIO.java
@@ -44,7 +44,7 @@ public class NBTIO {
             return Item.get(0);
         }
 
-        Item item = Item.get(tag.getShort("id"), (int) (!tag.contains("Damage") ? 0 : tag.getShort("Damage")), tag.getByte("Count"));
+        Item item = Item.get(tag.getShort("id"), !tag.contains("Damage") ? 0 : tag.getShort("Damage"), tag.getByte("Count"));
 
         if (tag.contains("tag") && tag.get("tag") instanceof CompoundTag) {
             item.setNamedTag(tag.getCompound("tag"));
