+++ src/main/java/cn/nukkit/tile/Chest.java
@@ -79,7 +79,7 @@ public class Chest extends Spawnable implements InventoryHolder, Container, Name
             return Item.get(Item.AIR, 0, 0);
         } else {
             CompoundTag data = (CompoundTag) this.namedTag.getList("Items").get(i);
-            return Item.get(data.getShort("id"), (int) data.getShort("Damage"), data.getByte("Count"));
+            return Item.get(data.getShort("id"), data.getShort("Damage"), data.getByte("Count"));
         }
     }
